benchmark/harness/requirements.txt
# Benchmark Harness Dependencies

# HTTP client
requests==2.31.0

# Metrics and statistics
hdrhistogram==0.10.3
numpy==1.24.3
pandas==2.0.3

# Load testing
locust==2.15.1

# Prometheus client
prometheus-client==0.17.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Progress bars
tqdm==4.66.1

# Configuration
pyyaml==6.0.1

# Date/time
python-dateutil==2.8.2

# JSON processing
jsonschema==4.19.0
